inherit sequencepairclassifier from sequenceclassifier
